requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
black==23.1.0
click==8.1.3
colorama==0.4.6
coverage==7.1.0
Django==4.1.6
django-extensions==3.2.1
django-htmx==1.13.0
django-taggit==3.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.1
flake8==6.0.0
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==6.2.5
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
